Use Config::AutoConf for analyzing libintl
The ad-hoc libintl detection code used in Makefile.PL was not taking
into account several of the flags that ExtUtil::MakeMaker can take
from the command line.

The following patch changes it to use Config::AutoConf in order to
prove for several libintl features and to generate the file
"config.h".